Change long function declaration wrapping.
I read about how a tool called Black does it and liked what I saw.

I don't think I'll be adding Black to the project but I'll be adopting a certain convention from it.

Full thing on one line -> if it's too long take the whole parenthesized part and drop it down two indent levels under the def statement (so it's not aligned with the function body) -> if that's still too long, every argument gets its own line.